[storage][geo] fix count(*) returns 0 bug.
It is caused by conversion between groogna's point in msec
and MySQL's point in degree. Groonga may return a point that
is not contained in degree (but it is conatined in msec) as
the first value. In the case, MySQL consider that all values
aren't included. So count(*) returns 0.

This commits checks a point returned by groonga is contained
in degree before the point is returned to MySQL.