Fix a crash bug when there is a syntax error query in many queries
A new grn_ctx is used for fulltext search because an allocated object
for fulltext search is destroyed after query is handled. It means that
handling a query and destroying an allocated object for fulltext
search may be occured in the same time on different threads. But
grn_ctx isn't thread safe. It may causes crash.

Reported by susumu takaoka. Thanks!!!