Add a test for a bug that matched record isn't deleted
fixes #1533

This problem is not related column type. It is caused when normal
index (not primary index) is used and

  (number of matched records / number of total records) < 0.5

. (The expression isn't correct but showes summary.)

The smallest case is 3 records and 1 match.

NOTE: This test finds another problem that deleted record is
selected. It may be groonga's problem. Groonga updates index wrongly
on deleting a record.