Stop to parse column comment in wrapper mode
We doen't need it. And it is backward incompatible.

It breaks database when unknown parameter error is occured in ALTER
TABLE. (It seems that this is a general wrapper mode bug. In wrapper
mode, we should do anything after validation. I didn't look at
code. So this may be not correct.)

Reported by Y.Kentaro. Thanks!!!