storage: support --auto-repair by mysqlcheck
It's an experimental feature.

If any data object such as table or column are broken (= lock is
remained), can't be repaired. If index columns are only broken (lock is
remained), Mroonga can recover the database.

Crashed check logic is just finding any existing lock. Because lock is
remained when Mroonga is crashed while updating. If other process opens
database for Mroonga, Mroonga may detect lock that is valid. You should
not use this feature when other process opens Mroonga database.