mariadb10.4 test: skip not suitable test
For not_equal.test, optimizer always uses "index" type not "range"
type with more records.

For not_found_in_limit.test, the expected "rows" is 2 not 1 with
MariaDB 10.4. Because the number of records in statistics is 2. It's
OK that "rows" is 1 or 2. We can add another test case that uses 2 as
the expected value or use --replace_result to accept 1 or 2 but there
are no reason to maintain this test case with additional cost.