ci linux: tests that always fail are disabled temporarily
Currently, these tests always fail.

The following tests are failing, but the cause of failure is the
Mroonga packages
for MariaDB10.6 don't exist in packages.groonga.org.
Because the Mroonga packages for MariaDB10.6 are made for the first
time.
Therefore, these tests disable temporarily.
We enable these tests again after we release the next release.