packages windows: add a patch for fixing build of Windows on CI
The download URL for FTP of PCRE2 can't use.
See: https://www.pcre.org/

Therefore, we need to use the URL of GitHub to download PCRE2 source code.

The cause of this problem is in MariaDB 10.5 or later.
We will remove this patch when this problem is resolved in MariaDB.

This problem has already been resolved at
https://github.com/MariaDB/server/commit/8d7196cdf165843b7b1271e2616bd57be359e1da

Therefore, this problem will resolve in the next release of MariaDB.