Fix max_execution_time test (#559)
`MAX_EXECUTION_TIME` didn't throw any error because of the MySQL's
`sleep()` function specification.


https://dev.mysql.com/doc/refman/8.0/en/miscellaneous-functions.html#function_sleep

> When
[SLEEP()](https://dev.mysql.com/doc/refman/8.0/en/miscellaneous-functions.html#function_sleep)
is the only thing invoked by a query that is interrupted, it returns 1
and the query itself returns no error. This is true whether the query is
killed or times out:

The current query seems to apply to this case.

> When
[SLEEP()](https://dev.mysql.com/doc/refman/8.0/en/miscellaneous-functions.html#function_sleep)
is only part of a query that is interrupted, the query returns an error:

I have changed the test query into like that.