packages mysql-8.0: remove installation of dependency packages of mysql-server
We build Mroonga package for Ubuntu on Launchpad,
but "apt build-dep -y mysql-server" command is not used on Launchpad
when we build Mroonga on Launchpad.
We use specification of Build-Depends in debian/control when we build
Mroonga on Launchpad.

Therefore, build of Mroonga package for Ubuntu failure on Launchpad
even if build of Mroonga package for Ubuntu successful on CI.

So, we remove "apt build-dep -y mysql-server" from CI.
We can notice missing dependency packages with CI by this PR.