test datetime fraction seconds: skip `1970-01-01 00:00:00.0` insert test on Windows (#794)
The following "out of range" error occurred only on Windows' MariaDB
when the datetime
value "1970-01-01 00:00:00.0" was set. So we decided to skip this test
on only Windows.

```
mysqltest: At line 34: query 'INSERT INTO diaries VALUES ("zero day", "1970-01-01 00:00:00.0")' failed: 1264: Out of range value for column 'created_at' at row 1
```

Note that it might be a valid value because it's between `1000-01-01
00:00:00.000000` - `9999-12-31 23:59:59.999999`:
https://mariadb.com/kb/en/datetime/#supported-values
But the error occurred.