bit: update index for NULL with `b'0'` to support index_column_diff
GitHub: GH-789

The current implementation ignores `NULL`. It means that we don't update index for `NULL`. But it causes `index_column_diff` false positive.
`BIT` with `NULL` is processed as `b'0'` in Groonga. Because Groonga uses the default value for `NULL`. `index_column_diff` uses `b'0'` for `NULL` in the expected posting lists. It causes false positive.

If we use `b'0'` instead of ignoring for `NULL`, we can avoid the false positive. It also enables that we can search `NULL` column value record by `b'0'`. In general, it'll be useful but this is an incompatible change. But it'll be acceptable because `NULL` behavior is undefined by definition.