timestamp: update index for `NULL` with `0000-00-00 00:00:00` to support `index_column_diff`
GitHub: GH-789

The current implementation ignores `NULL`. It means that we don't update index for `NULL`. But it causes `index_column_diff` false positive.

`timestamp` with `NULL` is processed as `0` (`0000-00-00 00:00:00` in MySQL/MariaDB) in Groonga. Because Groonga uses the default value for `NULL`. `index_column_diff` uses `0` for `NULL` in the expected posting lists. It causes false positive.

If we use `0`( `0000-00-00 00:00:00`  in MySQL/MariaDB) instead of ignoring for `NULL`, we can avoid the false positive. It also enables that we can search `NULL` column value record by `0000-00-00 00:00:00`. In general, it'll be useful but this is an incompatible change. But it'll be acceptable because NULL behavior is undefined by definition.

Note: MySQL enables STRICT_ALL_TABLES by default and it rejects `0000-00-00 00:00:00.0`.