storage bit: add support for bit
GitHub: fix GH-798

`BIT` support was entirely broken... So nobody uses it. This is
incompatible change but it has no problem.

We use the best fit size type instead of always using `UInt64`. For
example, we use `UInt8` for `BIT(4)` and `UInt32` for `BIT(24)`.

Index search is also supported by this.