char: use padded empty string for `NULL` (#810)
GitHub: GH-789

Mroonga doesn't support `NULL`. So inserting `NULL` to Mroonga table is
undefined by definition.

This change uses padded empty string for `NULL` for convenience. The
current behavior just uses the default value, an empty (not padded)
string. This is an incompatible change. But we can do it because this
behavior was undefined. Users who depend on this behavior must not
exist.