ci mariadb: add workflow to build Mroonga with MariaDB main
The purpose of this workflow is detect problem in
Mroonga that is bundled MariaDB in advance.

This change adds a new workflow that builds the latest
MariaDB main branch together with Mroonga. The goal is
to always use the most up-to-date Mroonga in our
MariaDB repository.

As a first step, we have configured the workflow to
build MariaDB main with a minimal set of plugins
(disabling unnecessary dynamic components) and to
check out the latest Mroonga code from its respective
repositories.

Co-authored-by: Horimoto Yasuhiro <horimoto@clear-code.com>