Setting encoding when we convert between() function (#999)
Fix: GH-984

This issue occurs because normalization fails inside
groonga-normalizer-mysql.
It happens when the `between()` function is condition-pushed down.

When the `between()` function is condition-pushed down, Mroonga converts
it into an internal representation so it can be processed internally.
However, during this conversion, the encoding is not set.
As a result, the converted `between()` function with the encoding
`GRN_ENC_NONE`.

Normalization in groonga-normalizer-mysql assumes that the encoding is
`GRN_ENC_UTF8`.
Therefore, when the encoding is `GRN_ENC_NONE`, normalization fails.

When condition push down occurs, we can fix this issue by applying that
encoding information inside Mroonga as well.
The encoding setting is provided by MariaDB.